[bbb-72] 🖋️Rename: NAVER API 호출 관련 패키지 외부로 이동
