--- src/main/java/com/bombombom/devs/book/naverapi/enums/SortType.java
@@ -1,6 +1,0 @@
-package com.bombombom.devs.book.naverapi.enums;
-
-public enum SortType {
-    SIM,
-    DATE
-}
