+++ src/main/java/com/bombombom/devs/client/naver/enums/SortType.java
@@ -1,0 +1,6 @@
+package com.bombombom.devs.client.naver.enums;
+
+public enum SortType {
+    SIM,
+    DATE
+}
