+++ src/main/java/com/bombombom/devs/client/naver/exception/ExternalApiException.java
@@ -1,4 +1,4 @@
-package com.bombombom.devs.book.naverapi.exception;
+package com.bombombom.devs.client.naver.exception;
 
 public class ExternalApiException extends RuntimeException {
