+++ src/main/java/com/bombombom/devs/global/exception/GlobalExceptionHandler.java
@@ -1,14 +1,12 @@
 package com.bombombom.devs.global.exception;
 
-import com.bombombom.devs.book.naverapi.exception.ExternalApiException;
+import com.bombombom.devs.client.naver.exception.ExternalApiException;
 import com.bombombom.devs.user.exception.ExistUsernameException;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.HashMap;
 import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
