+++ src/test/java/com/bombombom/devs/book/service/BookServiceTest.java
@@ -9,7 +9,6 @@ import static org.mockito.Mockito.verify;
 
 import com.bombombom.devs.book.enums.SearchOption;
 import com.bombombom.devs.book.models.Book;
-import com.bombombom.devs.book.naverapi.NaverClient;
 import com.bombombom.devs.book.repository.BookBulkRepository;
 import com.bombombom.devs.book.repository.BookRepository;
 import com.bombombom.devs.book.service.dto.AddBookCommand;
@@ -19,6 +18,7 @@ import com.bombombom.devs.book.service.dto.NaverBookApiResult.SearchBookItem;
 import com.bombombom.devs.book.service.dto.SearchBookQuery;
 import com.bombombom.devs.book.service.dto.SearchBooksResult;
 import com.bombombom.devs.book.service.dto.SearchBooksResult.BookResult;
+import com.bombombom.devs.client.naver.NaverClient;
 import java.util.Date;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
